+++ src/snowflake/snowpark/modin/plugin/docstrings/window.py
@@ -9,8 +9,8 @@ from textwrap import dedent
 
 from pandas.util._decorators import doc
 
-_rolling_agg_method_engine_template = """
-Compute the rolling {fname}.
+_window_agg_method_engine_template = """
+Compute the {win_type} {fname}.
 
 Parameters
 ----------
@@ -35,13 +35,12 @@ engine_kwargs : dict, default None {ek}
 
     **This parameter is ignored in Snowpark pandas. The execution engine will always be Snowflake.**
 
-**kwargs
-    Keyword arguments to be passed into func.
+{kwargs}
 
 Returns
 -------
 :class:`~snowflake.snowpark.modin.pandas.Series` or :class:`~snowflake.snowpark.modin.pandas.DataFrame`
-    Computed rolling {fname} of values.
+    Computed {win_type} {fname} of values.
 
 Examples
 --------
@@ -61,10 +60,10 @@ func : function, str, list, or dict
     - list of functions and/or function names, e.g. ``[np.sum, 'mean']``
     - dict of axis labels -> functions, function names or list of such.
 
-*args
+*args : tuple
     Positional arguments to pass to func.
 
-**kwargs
+**kwargs : dict
     Keyword arguments to be passed into func.
 
 Returns
@@ -112,17 +111,76 @@ Examples
 
 
 class Rolling:
+
+    """
+    Compute the rolling count.
+
+    Parameters
+    ----------
+    numeric_only : bool, default False
+        Include only float, int, boolean columns.
+
+    *args : tuple
+        Positional arguments to pass to func.
+
+    **kwargs : dict
+        Keyword arguments to be passed into func.
+
+    Returns
+    -------
+    :class:`~snowflake.snowpark.modin.pandas.Series` or :class:`~snowflake.snowpark.modin.pandas.DataFrame`
+        Computed rolling count of values.
+
+    Examples
+    --------
+    >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+    >>> df
+         B
+    0  0.0
+    1  1.0
+    2  2.0
+    3  NaN
+    4  4.0
+    >>> df.rolling(2, min_periods=1).count()
+       B
+    0  1
+    1  2
+    2  2
+    3  1
+    4  1
+    >>> df.rolling(2, min_periods=2).count()
+         B
+    0  NaN
+    1  2.0
+    2  2.0
+    3  1.0
+    4  1.0
+    >>> df.rolling(3, min_periods=1, center=True).count()
+       B
+    0  2
+    1  3
+    2  2
+    3  2
+    4  1
+    """
+
     def count():
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="sum",
         no=False,
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         e=None,
         ek=None,
@@ -163,12 +221,18 @@ class Rolling:
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="mean",
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         no=False,
         e=None,
@@ -211,9 +275,15 @@ class Rolling:
 
     # TODO: SNOW-1419071 API not implemented - uncomment when done.
     # @doc(
-    #     _rolling_agg_method_engine_template,
+    #     _window_agg_method_engine_template,
+    #     win_type="rolling",
     #     fname="median",
     #     args=None,
+    #     kwargs=dedent(
+    #         """\
+    #     **kwargs : dict
+    #         Keyword arguments to be passed into func."""
+    #     ),
     #     no=False,
     #     e=None,
     #     ek=None,
@@ -240,12 +310,18 @@ class Rolling:
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="var",
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         no=False,
         e=None,
@@ -287,12 +363,18 @@ class Rolling:
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="std",
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         no=False,
         e=None,
@@ -334,12 +416,18 @@ class Rolling:
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="min",
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         no=False,
         e=None,
@@ -367,12 +455,18 @@ class Rolling:
         pass
 
     @doc(
-        _rolling_agg_method_engine_template,
+        _window_agg_method_engine_template,
+        win_type="rolling",
         fname="max",
         args=dedent(
             """\
-        *args
+        *args : tuple
             Positional arguments to pass to func."""
+        ),
+        kwargs=dedent(
+            """\
+        **kwargs : dict
+            Keyword arguments to be passed into func."""
         ),
         no=False,
         e=None,
@@ -434,27 +528,228 @@ class Rolling:
 
 
 class Expanding:
+
+    """
+    Compute the expanding count.
+
+    Parameters
+    ----------
+    numeric_only : bool, default False
+        Include only float, int, boolean columns.
+
+    Returns
+    -------
+    :class:`~snowflake.snowpark.modin.pandas.Series` or :class:`~snowflake.snowpark.modin.pandas.DataFrame`
+        Computed expanding count of values.
+
+    Examples
+    --------
+    >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+    >>> df
+         B
+    0  0.0
+    1  1.0
+    2  2.0
+    3  NaN
+    4  4.0
+    >>> df.expanding(2).count()
+         B
+    0  NaN
+    1  2.0
+    2  3.0
+    3  3.0
+    4  4.0
+    """
+
     def count():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="sum",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).sum()
+             B
+        0  NaN
+        1  1.0
+        2  3.0
+        3  3.0
+        4  7.0"""
+        ),
+    )
     def sum():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="mean",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).mean()
+              B
+        0   NaN
+        1  0.50
+        2  1.00
+        3  1.00
+        4  1.75"""
+        ),
+    )
     def mean():
         pass
 
     def median():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="var",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).var()
+                  B
+        0       NaN
+        1  0.500000
+        2  1.000000
+        3  1.000000
+        4  2.916667"""
+        ),
+    )
     def var():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="std",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).std()
+                  B
+        0       NaN
+        1  0.707107
+        2  1.000000
+        3  1.000000
+        4  1.707825"""
+        ),
+    )
     def std():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="std",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).min()
+             B
+        0  NaN
+        1  0.0
+        2  0.0
+        3  0.0
+        4  0.0"""
+        ),
+    )
     def min():
         pass
 
+    @doc(
+        _window_agg_method_engine_template,
+        win_type="expanding",
+        fname="std",
+        no=False,
+        args=None,
+        kwargs=None,
+        e=None,
+        ek=None,
+        example=dedent(
+            """\
+        >>> df = pd.DataFrame({'B': [0, 1, 2, np.nan, 4]})
+        >>> df
+             B
+        0  0.0
+        1  1.0
+        2  2.0
+        3  NaN
+        4  4.0
+        >>> df.expanding(2).max()
+             B
+        0  NaN
+        1  1.0
+        2  2.0
+        3  2.0
+        4  4.0"""
+        ),
+    )
     def max():
         pass